+++ bot/twitter.py
@@ -1,66 +1,42 @@
 #!/usr/bin/env python3
-# -*- coding: utf-8 -*-
 """Methods for creating the stickers."""
 import datetime as dtm
 from io import BytesIO
+from typing import Union, cast, Optional
 
 import pytz
-from typing import Union
 from telegram import User
-from telegram.ext import CallbackContext
 from PIL import Image, ImageDraw, ImageFont, ImageFilter
-from hyphen import Hyphenator
 from hyphen.textwrap2 import fill
 
-TEMPLATE_DIRECTORY = 'templates'
-""":obj:`str`: Name of the directory containing the templates."""
-HEADER_TEMPLATE = f'{TEMPLATE_DIRECTORY}/header.png'
-""":obj:`str`: Path of the template for the header."""
-FOOTER_TEMPLATE = f'{TEMPLATE_DIRECTORY}/footer.png'
-""":obj:`str`: Path of the template for the footer."""
-BODY_TEMPLATE = f'{TEMPLATE_DIRECTORY}/body.png'
-""":obj:`str`: Path of the template for the body."""
-VERIFIED_TEMPLATE = f'{TEMPLATE_DIRECTORY}/verified.png'
-""":obj:`str`: Path of the template for the »verified« symbol."""
-VERIFIED_IMAGE = Image.open(VERIFIED_TEMPLATE)
-VERIFIED_IMAGE.thumbnail((27, 27))
-""":class:`Pillow.Image.Image`: The »verified« symbol as Pillow image in the correct size."""
-BACKGROUND = '#16202cff'
-""":obj:`str`: Background color."""
-TEXT_MAIN = '#ffffff'
-""":obj:`str`: Color of the main text."""
-TEXT_SECONDARY = '#8d99a5ff'
-""":obj:`str`: Color of secondary text."""
-FONTS_DIRECTORY = 'fonts'
-""":obj:`str`: Name of the directory containing the fonts."""
-FONT_HEAVY = f'{FONTS_DIRECTORY}/seguibl.ttf'
-""":obj:`str`: Font of the main text."""
-FONT_SEMI_BOLD = f'{FONTS_DIRECTORY}/seguisb.ttf'
-""":obj:`str`: Font of the secondary text."""
-FALLBACK_PROFILE_PICTURE = 'logo/TwitterStatusBot-rectangle.png'
-""":obj:`str`: Path of the picture to use as profile picture, if the user has none."""
-HEADERS_DIRECTORY = 'headers'
-""":obj:`str`: Name of the directory containing the saved headers."""
-FOOTER_FONT = ImageFont.truetype(FONT_SEMI_BOLD, 24)
-""":class:`PIL.ImageFont.Font`: Font to use for the footer."""
-USER_NAME_FONT = ImageFont.truetype(FONT_HEAVY, 24)
-""":class:`PIL.ImageFont.Font`: Font to use for the username."""
-USER_HANDLE_FONT = ImageFont.truetype(FONT_SEMI_BOLD, 23)
-""":class:`PIL.ImageFont.Font`: Font to use for the user handle."""
-BIG_TEXT_FONT = ImageFont.truetype(FONT_SEMI_BOLD, 70)
-""":class:`PIL.ImageFont.Font`: Font to use for big text in the body."""
-SMALL_TEXT_FONT = ImageFont.truetype(FONT_SEMI_BOLD, 36)
-""":class:`PIL.ImageFont.Font`: Font to use for small text in the body."""
-HYPHENATOR = Hyphenator('en_US')
-""":class:`PyHyphen.Hyphenator`: A hyphenator to use to wrap text."""
+from bot.constants import (
+    HEADER_TEMPLATE,
+    FOOTER_TEMPLATE,
+    BODY_TEMPLATE,
+    VERIFIED_IMAGE,
+    TEXT_MAIN,
+    TEXT_SECONDARY,
+    FALLBACK_PROFILE_PICTURE,
+    HEADERS_DIRECTORY,
+    FOOTER_FONT,
+    USER_NAME_FONT,
+    USER_HANDLE_FONT,
+    BIG_TEXT_FONT,
+    SMALL_TEXT_FONT,
+    HYPHENATOR,
+)
+from bot.userdata import CCT, UserData
 
 
 class HyphenationError(Exception):
+    """Custom exception class for hyphenation exceptions."""
 
     def __init__(self) -> None:
-        super().__init__('Something went wrong trying to hyphenate your text. Please note that '
-                         'words may not be longer than 100 characters. Also, currently only '
-                         'English is properly supported for hyphenation.')
+        super().__init__(
+            "Something went wrong trying to hyphenate your text. Please note that "
+            "words may not be longer than 100 characters. Also, currently only "
+            "English is properly supported for hyphenation."
+        )
 
 
 def mask_circle_transparent(image: Union[Image.Image, str]) -> Image.Image:
@@ -105,12 +81,12 @@ def shorten_text(text: str, max_width: int, font: ImageFont.ImageFont) -> str:
     short_text = text
     while width > max_width:
         i += 1
-        short_text = f'{text[:-i]}...'
+        short_text = f"{text[:-i]}..."
         (width, _), _ = font.font.getsize(short_text)
     return short_text
 
 
-def build_footer(timezone: str = 'Europe/Berlin') -> Image.Image:
+def build_footer(timezone: str = "Europe/Berlin") -> Image.Image:
     """
     Creates the footer for the sticker by adding the current timestamp.
 
@@ -122,7 +98,7 @@ def build_footer(timezone: str = 'Europe/Berlin') -> Image.Image:
         :class:`PIL.Image.Image`: The footer as Pillow image.
     """
     now = dtm.datetime.now(tz=pytz.timezone(timezone))
-    date_string = ' '.join([now.strftime('%I:%M %p'), '•', now.strftime('%b %d, %Y')])
+    date_string = " ".join([now.strftime("%I:%M %p"), "•", now.strftime("%b %d, %Y")])
 
     # Offsets
     top = 28
@@ -134,13 +110,15 @@ def build_footer(timezone: str = 'Europe/Berlin') -> Image.Image:
     return image
 
 
-def build_header(user: User, user_picture: Image.Image = None) -> Image.Image:
+def build_header(  # pylint: disable=R0914
+    user_data: UserData, user_picture: Image.Image = None
+) -> Image.Image:
     """
     Creates the header for the sticker customized for the given user. The header will be saved as
     file and can be reused.
 
     Args:
-        user: The Telegram user this header is build for.
+        user_data: The user data for the user this header is build for.
         user_picture: Optional. The profile picture of the user. Defaults to the bots logo.
 
     Returns:
@@ -157,6 +135,14 @@ def build_header(user: User, user_picture: Image.Image = None) -> Image.Image:
     if not user_picture:
         user_picture = mask_circle_transparent(FALLBACK_PROFILE_PICTURE)
 
+    # crop a centered square
+    if not user_picture.width == user_picture.height:
+        side = min(user_picture.width, user_picture.height)
+        left = (user_picture.width - side) // 2
+        upper = (user_picture.height - side) // 2
+        user_picture = user_picture.crop((left, upper, left + side, upper + side))
+
+    # make it a circle an small
     user_picture = mask_circle_transparent(user_picture)
     user_picture.thumbnail((78, 78))
     background.alpha_composite(user_picture, (up_left, up_top))
@@ -165,14 +151,16 @@ def build_header(user: User, user_picture: Image.Image = None) -> Image.Image:
     un_left = 118
     un_top = 30
     draw = ImageDraw.Draw(background)
-    user_name = shorten_text(user.full_name, 314, USER_NAME_FONT)
+    user_name = shorten_text(cast(str, user_data.full_name), 314, USER_NAME_FONT)
     draw.text((un_left, un_top), user_name, fill=TEXT_MAIN, font=USER_NAME_FONT)
 
     # Add user handle
     uh_left = 118
     uh_top = 62
     draw = ImageDraw.Draw(background)
-    user_handle = shorten_text(f'@{user.username or user.first_name}', 370, USER_HANDLE_FONT)
+    user_handle = shorten_text(
+        f"@{user_data.username or user_data.first_name}", 370, USER_HANDLE_FONT
+    )
     draw.text((uh_left, uh_top), user_handle, fill=TEXT_SECONDARY, font=USER_HANDLE_FONT)
 
     # Add verified symbol
@@ -182,14 +170,16 @@ def build_header(user: User, user_picture: Image.Image = None) -> Image.Image:
     background.alpha_composite(VERIFIED_IMAGE, (v_left, v_top))
 
     # Save for later use
-    background.save(f'{HEADERS_DIRECTORY}/{user.id}.png')
+    background.save(f"{HEADERS_DIRECTORY}/{user_data.user_id}.png")
     return background
 
 
-def get_header(user: User, context: CallbackContext) -> Image.Image:
+def get_header(user: User, context: CCT) -> Image.Image:
     """
     Gets the header for the sticker customized for the given user. The header either be loaded from
     file or created anew, if there is no header for the user or users info changed.
+
+    If the header needs to be generated anew, stored stickers are deleted.
 
     Args:
         user: The Telegram user this header is build for.
@@ -200,32 +190,43 @@ def get_header(user: User, context: CallbackContext) -> Image.Image:
     Returns:
         :class:`PIL.Image.Image`: The header as Pillow image.
     """
+    drop_stored_stickers = False
     bot = context.bot
+    user_data = cast(UserData, context.user_data)
 
     profile_photos = bot.get_user_profile_photos(user.id, limit=1)
-    if profile_photos.total_count > 0:
+    if profile_photos and profile_photos.total_count > 0:
         photo = profile_photos.photos[0][0]
-        photo_file_id = photo.file_id
-        photo_file_unique_id = photo.file_unique_id
+        photo_file_id: Optional[str] = photo.file_id
+        photo_file_unique_id: Optional[str] = photo.file_unique_id
     else:
         photo_file_id = None
         photo_file_unique_id = None
-    user.photo_file_unique_id = photo_file_unique_id
 
-    stored_user = context.bot_data.get(user.id, None)
-    if (stored_user and stored_user.full_name == user.full_name
-            and stored_user.username == user.username
-            and stored_user.photo_file_unique_id == user.photo_file_unique_id):
+    fallback_file_id = user_data.fallback_photo.file_id if user_data.fallback_photo else None
+    fallback_unique_id = (
+        user_data.fallback_photo.file_unique_id if user_data.fallback_photo else None
+    )
+
+    if (
+        user_data.full_name == user.full_name  # pylint: disable=R0916
+        and user_data.username == user.username
+        and user_data.photo_file_unique_id == (photo_file_unique_id or fallback_unique_id)
+    ):
         # Try to return saved header
-        # If saving failed, we need to create a new one
         try:
-            return Image.open(f'{HEADERS_DIRECTORY}/{user.id}.png')
+            return Image.open(f"{HEADERS_DIRECTORY}/{user.id}.png")
         except FileNotFoundError:
+            # If saving failed, we need to create a new one
             pass
+    else:
+        drop_stored_stickers = True
 
     # Get users profile picture, if it exists
-    if photo_file_id:
-        photo_file = bot.get_file(photo_file_id)
+    file_id = photo_file_id or fallback_file_id
+    file_unique_id = photo_file_unique_id or fallback_unique_id
+    if file_id:
+        photo_file = bot.get_file(file_id)
         picture_stream = BytesIO()
         photo_file.download(out=picture_stream)
         picture_stream.seek(0)
@@ -233,8 +234,10 @@ def get_header(user: User, context: CallbackContext) -> Image.Image:
     else:
         user_picture = None
 
-    context.bot_data[user.id] = user
-    return build_header(user, user_picture)
+    user_data.update_user_info(user, photo_file_unique_id=file_unique_id)
+    if drop_stored_stickers:
+        user_data.sticker_file_ids.clear()
+    return build_header(user_data, user_picture)
 
 
 def build_body(text: str) -> Image.Image:
@@ -250,34 +253,38 @@ def build_body(text: str) -> Image.Image:
     max_chars_per_line = 26
     max_pixels_per_line = 450
 
-    def single_line_text(position, text_, font, bg):  # type: ignore
+    def single_line_text(position, text_, font, background_):  # type: ignore
         _, height = font.getsize(text_)
-        bg = bg.resize((bg.width, height + top + 1))
-        draw = ImageDraw.Draw(bg)
+        background_ = background_.resize((background_.width, height + top + 1))
+        draw = ImageDraw.Draw(background_)
         draw.text(position, text_, fill=TEXT_MAIN, font=font)
 
-        return bg
+        return background_
 
-    def multiline_text(position, text_, bg):  # type: ignore
+    def multiline_text(position, text_, background__):  # type: ignore
         spacing = 4
         _, height = SMALL_TEXT_FONT.getsize_multiline(text_, spacing=spacing)
-        bg = bg.resize((bg.width, height - spacing))
-        draw = ImageDraw.Draw(bg)
-        draw.multiline_text(position, text_, fill=TEXT_MAIN, font=SMALL_TEXT_FONT, spacing=spacing)
+        background__ = background__.resize((background__.width, height - spacing))
+        draw = ImageDraw.Draw(background__)
+        draw.multiline_text(
+            position,
+            text_,
+        )
 
-        return bg
+        return background__
 
     background = Image.open(BODY_TEMPLATE)
     left = 27
 
-    if '\n' in text:
+    if "\n" in text:
         top = -12
-        lines = text.split('\n')
+        lines = text.split("\n")
         try:
-            text = '\n'.join(
-                [fill(line, max_chars_per_line, use_hyphenator=HYPHENATOR) for line in lines])
-        except Exception:
-            raise HyphenationError
+            text = "\n".join(
+                [fill(line, max_chars_per_line, use_hyphenator=HYPHENATOR) for line in lines]
+            )
+        except Exception as exc:
+            raise HyphenationError from exc
         background = multiline_text((left, top), text, background)
     else:
         width, _ = BIG_TEXT_FONT.getsize(text)
@@ -287,8 +294,8 @@ def build_body(text: str) -> Image.Image:
             if width > max_pixels_per_line:
                 try:
                     text = fill(text, max_chars_per_line, use_hyphenator=HYPHENATOR)
-                except Exception:
-                    raise HyphenationError
+                except Exception as exc:
+                    raise HyphenationError from exc
                 background = multiline_text((left, top), text, background)
             else:
                 background = single_line_text((left, top), text, SMALL_TEXT_FONT, background)
@@ -299,12 +306,19 @@ def build_body(text: str) -> Image.Image:
     return background
 
 
-def build_sticker(text: str, user: User, context: CallbackContext) -> Image.Image:
+def build_sticker(text: str, user: User, context: CCT) -> Image.Image:
+    """Builds the sticker.
+
+    Arguments:
+        text: Text of the tweet.
+        user: The user the sticker is generated for.
+        context: The callback context as provided by the dispatcher.
+    """
     header = get_header(user, context)
     body = build_body(text)
     footer = build_footer()
 
-    sticker = Image.new('RGBA', (512, header.height + body.height + footer.height))
+    sticker = Image.new("RGBA", (512, header.height + body.height + footer.height))
     sticker.paste(header, (0, 0))
     sticker.paste(body, (0, header.height))
     sticker.paste(footer, (0, header.height + body.height))