+++ main.py
@@ -1,19 +1,23 @@
 #!/usr/bin/env python3
-# -*- coding: utf-8 -*-
 """The script that runs the bot."""
 import logging
 from configparser import ConfigParser
+from typing import cast, Dict
+
 from telegram import ParseMode
-from telegram.ext import Updater, Defaults, PicklePersistence, Filters
+from telegram.ext import Updater, Defaults, PicklePersistence, ContextTypes
 
-from ptbstats import set_dispatcher, register_stats, SimpleStats
-
-import bot
+from bot.setup import setup_dispatcher
+from bot.userdata import UserData, CCT
 
 # Enable logging
-logging.basicConfig(format='%(asctime)s - %(name)s - %(levelname)s - %(message)s',
-                    level=logging.INFO,
-                    filename='tsb.log')
+logging.basicConfig(
+    format="%(asctime)s - %(name)s - %(levelname)s - %(message)s",
+    level=logging.INFO,
+    # filename="tsb.log",
+)
+aps_logger = logging.getLogger('apscheduler')
+aps_logger.setLevel(logging.WARNING)
 
 logger = logging.getLogger(__name__)
 
@@ -22,33 +26,26 @@ def main() -> None:
     """Start the bot."""
     # Read configuration values from bot.ini
     config = ConfigParser()
-    config.read('bot.ini')
-    token = config['TwitterStatusBot']['token']
-    admin = int(config['TwitterStatusBot']['admins_chat_id'])
+    config.read("bot.ini")
+    token = config["TwitterStatusBot"]["token"]
+    admin_id = int(config["TwitterStatusBot"]["admins_chat_id"])
+    sticker_set_name = config["TwitterStatusBot"]["sticker_set_name"]
 
-    # Create the Updater and pass it your bot's token.
-    # Make sure to set use_context=True to use the new context based callbacks
-    # Post version 12 this will no longer be necessary
-    defaults = Defaults(parse_mode=ParseMode.HTML, disable_notification=True)
-    persistence = PicklePersistence('tsb.pickle')
-    updater = Updater(token,
-                      use_context=True,
-                      defaults=defaults,
-                      persistence=persistence,
-                      workers=8)
-
-    # Set up stats
-    set_dispatcher(updater.dispatcher)
-    register_stats(SimpleStats('ilq', lambda u: bool(u.inline_query and u.inline_query.query)),
-                   admin_id=admin)
-    register_stats(SimpleStats(
-        'text', lambda u: bool(u.effective_message and
-                               (Filters.chat_type.private & Filters.text & ~Filters.command)(u))),
-                   admin_id=admin)
+    defaults = Defaults(parse_mode=ParseMode.HTML, disable_notification=True, run_async=True)
+    # Cast due to a type hinting bug in ptb v13.6
+    context_types = cast(ContextTypes[CCT, UserData, Dict, Dict], ContextTypes(user_data=UserData))
+    persistence = PicklePersistence("tsb.pickle", context_types=context_types, single_file=False)
+    updater = Updater(
+        token,
+        defaults=defaults,
+        persistence=persistence,
+        workers=8,
+        context_types=context_types,
+    )
 
     # Get the dispatcher to register handlers
-    dp = updater.dispatcher
-    bot.register_dispatcher(dp)
+    dispatcher = updater.dispatcher
+    setup_dispatcher(dispatcher, admin_id, sticker_set_name)
 
     # Start the Bot
     updater.start_polling()
@@ -59,5 +56,5 @@ def main() -> None:
     updater.idle()
 
 
-if __name__ == '__main__':
+if __name__ == "__main__":
     main()